Move models to own namespace
