+++ grant/models/__init__.py
@@ -1,0 +1,5 @@
+""" Models for use in windows """
+
+from .table_model import TableModel
+from .tasks_model import TasksModel
+from .tree_model import TreeModel
