+++ tests/test_tree_model.py
@@ -4,8 +4,8 @@ from PyQt5.QtCore import QModelIndex
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QIcon
 
-from grant.windows.tree_model import TreeNode
-from grant.windows.tree_model import TreeModel
+from grant.models.tree_model import TreeNode
+from grant.models.tree_model import TreeModel
 from grant.research import ResearchProject, ResearchPlan, ResearchTask
